upload option in download_figs, fixed openmp import conflicts

+++ download_figures.sh
@@ -3,6 +3,15 @@ REMOTE_HOST="engr-mahala01s.engr.arizona.edu"
 REMOTE_USER="berian"
 REMOTE_BASE="/workspace/berian/shapenet_sar_sim"
 
+UPLOAD=0
+while getopts "u" opt; do
+    case "$opt" in
+        u) UPLOAD=1 ;;
+        *) echo "Usage: $0 [-u] [filter]" >&2; exit 1 ;;
+    esac
+done
+shift $((OPTIND - 1))
+
 # Each entry is "remote_subdir:local_path"
 FOLDERS=(
     "figures:/home/berian/Documents/remote-figures"
@@ -17,7 +26,12 @@ fi
 for entry in "${FOLDERS[@]}"; do
     REMOTE_SUBDIR="${entry%%:*}"
     LOCAL_PATH="${entry#*:}"
-    rm -rf "$LOCAL_PATH"/*
-    mkdir -p "$LOCAL_PATH"
-    rsync "${RSYNC_ARGS[@]}" "$REMOTE_USER@$REMOTE_HOST:$REMOTE_BASE/$REMOTE_SUBDIR/" "$LOCAL_PATH/"
+    REMOTE="$REMOTE_USER@$REMOTE_HOST:$REMOTE_BASE/$REMOTE_SUBDIR/"
+    if [ "$UPLOAD" -eq 1 ]; then
+        rsync "${RSYNC_ARGS[@]}" "$LOCAL_PATH/" "$REMOTE"
+    else
+        rm -rf "$LOCAL_PATH"/*
+        mkdir -p "$LOCAL_PATH"
+        rsync "${RSYNC_ARGS[@]}" "$REMOTE" "$LOCAL_PATH/"
+    fi
 done
